reorganized api files to improve readability
