+++ app/pages/handlers.go
@@ -1,20 +1,19 @@
-package main
+package pages
 
 import (
 	"net/http"
 
-	"github.com/demostanis/hypertube/pages"
 	. "maragu.dev/gomponents"
 )
 
 func HomeHandler(w http.ResponseWriter, r *http.Request) (Node, error) {
-	return pages.Home(), nil
+	return Home(), nil
 }
 
 func LoginHandler(w http.ResponseWriter, r *http.Request) (Node, error) {
-	return pages.Login(), nil
+	return Login(), nil
 }
 
 func SigninHandler(w http.ResponseWriter, r *http.Request) (Node, error) {
-	return pages.Signin(), nil
+	return Signin(), nil
 }
